LayoutTests/http/tests/digital-credentials/digital-credential-console-messages.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Digital Credential API: Console Message Tests</title>
<script src="/js-test-resources/js-test.js"></script>
<body onload="setup().then(generateConsoleMessages)"></body>
<script>
    async function setup() {
        if (document.visibilityState === "hidden") {
            await new Promise((resolve) => {
                document.onvisibilitychange = resolve;
                testRunner.setPageVisibility("visible");
            });
        }
    }

    async function generateConsoleMessages() {
        try {
            await navigator.credentials.get({
                digital: {
                    requests: [
                        { protocol: "unknown-protocol-1", data: {} },
                        { protocol: "unknown-protocol-2", data: {} },
                    ],
                },
                mediation: "required",
            });
        } catch {}

        try {
            await navigator.credentials.get({
                digital: {
                    requests: [
                        { protocol: "unknown-before", data: {} },
                        {
                            protocol: "org-iso-mdoc",
                            data: {
                                deviceRequest:
                                    "omd2ZXJzaW9uYzEuMGtkb2NSZXF1ZXN0c4GhbGl0ZW1zUmVxdWVzdNgYWIKiZ2RvY1R5cGV1b3JnLmlzby4xODAxMy41LjEubURMam5hbWVTcGFjZXOhcW9yZy5pc28uMTgwMTMuNS4x9pWthZ2Vfb3Zlcl8yMfRqZ2l2ZW5fbmFtZfRrZmFtaWx5X25hbWX0cmRyaXZpbmdfcHJpdmlsZWdlc_RocG9ydHJhaXT0",
                                encryptionInfo:
                                    "gmVkY2FwaaJlbm9uY2VYICBetSsDkKlE_G9JSIHwPzr3ctt6Ol9GgmCH8iGdGQNJcnJlY2lwaWVudFB1YmxpY0tleaQBAiABIVggKKm1iPeuOb9bDJeeJEL4QldYlWvY7F_K8eZkmYdS9PwiWCCm9PLEmosiE_ildsE11lqq4kDkjhfQUKPpbX-Hm1ZSLg",
                            },
                        },
                        { protocol: "unknown-after", data: {} },
                    ],
                },
                mediation: "required",
            });
        } catch {}
    }
</script>
