LayoutTests/http/tests/digital-credentials/digital-credential-console-messages.https-expected.txt

CONSOLE MESSAGE: Ignoring DigitalCredentialGetRequest with unsupported protocol: "unknown-protocol-1"
CONSOLE MESSAGE: Ignoring DigitalCredentialGetRequest with unsupported protocol: "unknown-protocol-2"
CONSOLE MESSAGE: Ignoring DigitalCredentialGetRequest with unsupported protocol: "unknown-before"
CONSOLE MESSAGE: Ignoring DigitalCredentialGetRequest with unsupported protocol: "unknown-after"
PASS successfullyParsed is true

TEST COMPLETE
